requirements.txt
albucore==0.0.21
albumentations==1.4.22
annotated-types==0.7.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
eval_type_backport==0.2.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
gdown==5.2.0
idna==3.10
imageio==2.36.1
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.0
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
simsimd==6.2.1
six==1.17.0
soupsieve==2.6
stringzilla==3.11.1
sympy==1.13.1
tifffile==2024.12.12
torch==2.6.0.dev20241215
torchaudio==2.6.0.dev20241215
torchvision==0.22.0.dev20241215
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
